Introduce RepublishService
At the moment it has the same behaviour of RepublishWorker, but means it
can be used outside of a Sidekiq context.
