+++ app/workers/republish_worker.rb
@@ -1,64 +1,7 @@
-require "services"
-
 class RepublishWorker
   include Sidekiq::Worker
 
   def perform(content_id)
-    document = Document.find(content_id)
-
-    if document.publication_state == "published"
-      document.update_type = "republish"
-
-      publishing_api_put_content(document)
-      publishing_api_patch_links(document)
-      publishing_api_publish(document)
-    elsif document.publication_state == "draft"
-      published_edition_version_number = document.state_history.key("published")
-
-      publishing_api_patch_links(document)
-
-      if published_edition_version_number.present?
-        published_document = Document.find(
-          content_id,
-          version: published_edition_version_number,
-        )
-        published_document.update_type = "republish"
-
-        publishing_api_put_content(published_document)
-        publishing_api_publish(published_document)
-      end
-
-      publishing_api_put_content(document)
-    else
-      print_limitations_of_republishing(document)
-    end
-  end
-
-private
-
-  def print_limitations_of_republishing(document)
-    content_id = document.content_id
-    state = document.publication_state
-
-    message = "Skipped republishing document with content_id #{content_id}"
-    message += " because it has a state of '#{state}'. Currently, there is"
-    message += " no way to safely republish the document that is supported by"
-    message += " the platform."
-
-    logger.warn message
-  end
-
-  def publishing_api_put_content(document)
-    payload = DocumentPresenter.new(document).to_json
-    Services.publishing_api.put_content(document.content_id, payload)
-  end
-
-  def publishing_api_patch_links(document)
-    payload = DocumentLinksPresenter.new(document).to_json
-    Services.publishing_api.patch_links(document.content_id, payload)
-  end
-
-  def publishing_api_publish(document)
-    Services.publishing_api.publish(document.content_id, "republish")
+    RepublishService.new.call(content_id)
   end
 end
